requirements.txt
Flask==1.0.2
PyPDF2==1.26.0
python-dateutil
requests==2.6.0
